# Downloads/gibifinder-flyio/gibifinder/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GibiFinder API - Comic book identifier using AI vision
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface IdentifyComicRequest {
  /**
   * Base64 encoded images
   * @minItems 1
   * @maxItems 3
   */
  images: string[];
  /** MIME types of the images (e.g. image/jpeg) */
  mimeTypes: string[];
}

export interface TextSearchRequest {
  /** Character name or comic title */
  query: string;
}

export interface DescriptionSearchRequest {
  /** Description of a comic scene or story */
  description: string;
}

export interface QuoteSearchRequest {
  /** Text from a speech bubble to search for */
  quote: string;
}

export interface HistoryItem {
  id: number;
  titulo: string;
  editora: string;
  personagens?: string[];
  descricao?: string;
  confianca: number;
  nota?: string;
  balloonText?: string;
  imageThumbnail?: string;
  createdAt?: string;
  averageRating?: number;
  totalRatings?: number;
  searchType?: string;
  anoLancamento?: string;
  numeroPagina?: string;
  rank?: number;
}

export interface QuoteSearchResponse {
  results: HistoryItem[];
  total: number;
}

export type ComicResultSearchType =
  (typeof ComicResultSearchType)[keyof typeof ComicResultSearchType];

export const ComicResultSearchType = {
  image: "image",
  text: "text",
  description: "description",
} as const;

export interface ComicResult {
  /** DB ID if saved */
  id?: number;
  encontrado: boolean;
  titulo: string;
  editora: string;
  personagens: string[];
  descricao: string;
  /**
   * @minimum 0
   * @maximum 100
   */
  confianca: number;
  nota: string;
  balloonText?: string;
  /** Base64 thumbnail of first image */
  imageThumbnail?: string;
  /** Average user rating (1=thumbs up, 0=thumbs down) */
  averageRating?: number;
  totalRatings?: number;
  searchType?: ComicResultSearchType;
  anoLancamento?: string;
  numeroPagina?: string;
  rank?: number;
}

export interface ComicSearchResponse {
  results: ComicResult[];
}

export interface ComicDetail {
  id: number;
  encontrado: boolean;
  titulo: string;
  editora: string;
  personagens: string[];
  descricao: string;
  confianca: number;
  nota: string;
  balloonText?: string;
  imageThumbnail?: string;
  createdAt?: string;
  averageRating?: number;
  totalRatings?: number;
  searchType?: string;
  anoLancamento?: string;
  numeroPagina?: string;
  rank?: number;
}

export interface HistoryResponse {
  items: HistoryItem[];
  total: number;
  page: number;
  limit: number;
}

export interface RankingItem {
  id?: number;
  titulo: string;
  editora: string;
  imageThumbnail?: string;
  searchCount: number;
  averageRating?: number;
  totalRatings?: number;
}

export interface RankingResponse {
  items: RankingItem[];
  weekStart: string;
  weekEnd: string;
}

/**
 * 0 = thumbs down, 1 = thumbs up
 */
export type RatingRequestRating =
  (typeof RatingRequestRating)[keyof typeof RatingRequestRating];

export const RatingRequestRating = {
  NUMBER_0: 0,
  NUMBER_1: 1,
} as const;

export interface RatingRequest {
  comicId: number;
  /** 0 = thumbs down, 1 = thumbs up */
  rating: RatingRequestRating;
  /** Text feedback when thumbs down */
  feedback?: string;
}

export interface RatingResponse {
  success: boolean;
  averageRating: number;
  totalRatings: number;
}

export type GetHistoryParams = {
  titulo?: string;
  editora?: string;
  page?: number;
  limit?: number;
};