# Downloads/gibifinder-flyio/gibifinder/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GibiFinder API - Comic book identifier using AI vision
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: comics
    description: Comic identification and search
  - name: history
    description: Search history
  - name: ranking
    description: Weekly ranking
  - name: ratings
    description: User ratings
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /comics/identify:
    post:
      operationId: identifyComic
      tags: [comics]
      summary: Identify comic from images
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdentifyComicRequest"
      responses:
        "200":
          description: Identification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComicResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /comics/search:
    post:
      operationId: searchComic
      tags: [comics]
      summary: Search comic by text (character name or title)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TextSearchRequest"
      responses:
        "200":
          description: Search result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComicResult"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /comics/quote-search:
    post:
      operationId: quoteSearch
      tags: [comics]
      summary: Search comic by balloon text quote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteSearchRequest"
      responses:
        "200":
          description: Search results by quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteSearchResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /comics/{id}:
    get:
      operationId: getComic
      tags: [comics]
      summary: Get comic by ID (shareable page)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comic details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComicDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /history:
    get:
      operationId: getHistory
      tags: [history]
      summary: Get search history
      parameters:
        - name: titulo
          in: query
          schema:
            type: string
        - name: editora
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: History list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryResponse"

  /ranking:
    get:
      operationId: getRanking
      tags: [ranking]
      summary: Get weekly ranking top 10
      responses:
        "200":
          description: Ranking list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RankingResponse"

  /ratings:
    post:
      operationId: submitRating
      tags: [ratings]
      summary: Submit a rating for a comic result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RatingRequest"
      responses:
        "200":
          description: Rating saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RatingResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message

    IdentifyComicRequest:
      type: object
      properties:
        images:
          type: array
          items:
            type: string
          minItems: 1
          maxItems: 3
          description: Base64 encoded images
        mimeTypes:
          type: array
          items:
            type: string
          description: MIME types of the images (e.g. image/jpeg)
      required:
        - images
        - mimeTypes

    TextSearchRequest:
      type: object
      properties:
        query:
          type: string
          description: Character name or comic title
      required:
        - query

    QuoteSearchRequest:
      type: object
      properties:
        quote:
          type: string
          description: Text from a speech bubble to search for
      required:
        - quote

    QuoteSearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/HistoryItem"
        total:
          type: integer
      required:
        - results
        - total

    ComicResult:
      type: object
      properties:
        id:
          type: integer
          description: DB ID if saved
        encontrado:
          type: boolean
        titulo:
          type: string
        editora:
          type: string
        personagens:
          type: array
          items:
            type: string
        descricao:
          type: string
        confianca:
          type: integer
          minimum: 0
          maximum: 100
        nota:
          type: string
        balloonText:
          type: string
        imageThumbnail:
          type: string
          description: Base64 thumbnail of first image
        averageRating:
          type: number
          description: Average user rating (1=thumbs up, 0=thumbs down)
        totalRatings:
          type: integer
        searchType:
          type: string
          enum: [image, text]
      required:
        - encontrado
        - titulo
        - editora
        - personagens
        - descricao
        - confianca
        - nota

    ComicDetail:
      type: object
      properties:
        id:
          type: integer
        encontrado:
          type: boolean
        titulo:
          type: string
        editora:
          type: string
        personagens:
          type: array
          items:
            type: string
        descricao:
          type: string
        confianca:
          type: integer
        nota:
          type: string
        balloonText:
          type: string
        imageThumbnail:
          type: string
        createdAt:
          type: string
        averageRating:
          type: number
        totalRatings:
          type: integer
        searchType:
          type: string
      required:
        - id
        - encontrado
        - titulo
        - editora
        - personagens
        - descricao
        - confianca
        - nota

    HistoryItem:
      type: object
      properties:
        id:
          type: integer
        titulo:
          type: string
        editora:
          type: string
        personagens:
          type: array
          items:
            type: string
        descricao:
          type: string
        confianca:
          type: integer
        nota:
          type: string
        balloonText:
          type: string
        imageThumbnail:
          type: string
        createdAt:
          type: string
        averageRating:
          type: number
        totalRatings:
          type: integer
        searchType:
          type: string
      required:
        - id
        - titulo
        - editora
        - confianca

    HistoryResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/HistoryItem"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - items
        - total
        - page
        - limit

    RankingItem:
      type: object
      properties:
        id:
          type: integer
        titulo:
          type: string
        editora:
          type: string
        imageThumbnail:
          type: string
        searchCount:
          type: integer
        averageRating:
          type: number
        totalRatings:
          type: integer
      required:
        - titulo
        - editora
        - searchCount

    RankingResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/RankingItem"
        weekStart:
          type: string
        weekEnd:
          type: string
      required:
        - items
        - weekStart
        - weekEnd

    RatingRequest:
      type: object
      properties:
        comicId:
          type: integer
        rating:
          type: integer
          enum: [0, 1]
          description: "0 = thumbs down, 1 = thumbs up"
      required:
        - comicId
        - rating

    RatingResponse:
      type: object
      properties:
        success:
          type: boolean
        averageRating:
          type: number
        totalRatings:
          type: integer
      required:
        - success
        - averageRating
        - totalRatings