# Downloads/gibifinder-flyio/gibifinder/lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GibiFinder API - Comic book identifier using AI vision
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Identify comic from images
 */
export const identifyComicBodyImagesMax = 3;

export const IdentifyComicBody = zod.object({
  images: zod
    .array(zod.string())
    .min(1)
    .max(identifyComicBodyImagesMax)
    .describe("Base64 encoded images"),
  mimeTypes: zod
    .array(zod.string())
    .describe("MIME types of the images (e.g. image\/jpeg)"),
});

const ComicResultItem = zod.object({
  id: zod.number().optional().describe("DB ID if saved"),
  encontrado: zod.boolean(),
  titulo: zod.string(),
  editora: zod.string(),
  personagens: zod.array(zod.string()),
  descricao: zod.string(),
  confianca: zod.number().min(0).max(100),
  nota: zod.string(),
  balloonText: zod.string().optional(),
  imageThumbnail: zod.string().optional(),
  averageRating: zod.number().optional(),
  totalRatings: zod.number().optional(),
  searchType: zod.enum(["image", "text", "description"]).optional(),
  anoLancamento: zod.string().optional(),
  numeroPagina: zod.string().optional(),
  rank: zod.number().optional(),
});

export const identifyComicResponseConfiancaMin = 0;
export const identifyComicResponseConfiancaMax = 100;

export const IdentifyComicResponse = zod.object({
  results: zod.array(ComicResultItem),
});

/**
 * @summary Search comic by text (character name or title)
 */
export const SearchComicBody = zod.object({
  query: zod.string().describe("Character name or comic title"),
});

export const searchComicResponseConfiancaMin = 0;
export const searchComicResponseConfiancaMax = 100;

export const SearchComicResponse = zod.object({
  results: zod.array(ComicResultItem),
});

/**
 * @summary Search comic by description
 */
export const DescriptionSearchBody = zod.object({
  description: zod.string().min(1).describe("Description of a comic scene or story"),
});

export const DescriptionSearchResponse = zod.object({
  results: zod.array(ComicResultItem),
});

/**
 * @summary Search comic by balloon text quote
 */
export const QuoteSearchBody = zod.object({
  quote: zod.string().describe("Text from a speech bubble to search for"),
});

export const QuoteSearchResponse = zod.object({
  results: zod.array(
    zod.object({
      id: zod.number(),
      titulo: zod.string(),
      editora: zod.string(),
      personagens: zod.array(zod.string()).optional(),
      descricao: zod.string().optional(),
      confianca: zod.number(),
      nota: zod.string().optional(),
      balloonText: zod.string().optional(),
      imageThumbnail: zod.string().optional(),
      createdAt: zod.string().optional(),
      averageRating: zod.number().optional(),
      totalRatings: zod.number().optional(),
      searchType: zod.string().optional(),
      anoLancamento: zod.string().optional(),
      numeroPagina: zod.string().optional(),
      rank: zod.number().optional(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Get comic by ID (shareable page)
 */
export const GetComicParams = zod.object({
  id: zod.coerce.number(),
});

export const GetComicResponse = zod.object({
  id: zod.number(),
  encontrado: zod.boolean(),
  titulo: zod.string(),
  editora: zod.string(),
  personagens: zod.array(zod.string()),
  descricao: zod.string(),
  confianca: zod.number(),
  nota: zod.string(),
  balloonText: zod.string().optional(),
  imageThumbnail: zod.string().optional(),
  createdAt: zod.string().optional(),
  averageRating: zod.number().optional(),
  totalRatings: zod.number().optional(),
  searchType: zod.string().optional(),
  anoLancamento: zod.string().optional(),
  numeroPagina: zod.string().optional(),
  rank: zod.number().optional(),
});

/**
 * @summary Get search history
 */
export const getHistoryQueryPageDefault = 1;
export const getHistoryQueryLimitDefault = 20;

export const GetHistoryQueryParams = zod.object({
  titulo: zod.coerce.string().optional(),
  editora: zod.coerce.string().optional(),
  page: zod.coerce.number().default(getHistoryQueryPageDefault),
  limit: zod.coerce.number().default(getHistoryQueryLimitDefault),
});

export const GetHistoryResponse = zod.object({
  items: zod.array(
    zod.object({
      id: zod.number(),
      titulo: zod.string(),
      editora: zod.string(),
      personagens: zod.array(zod.string()).optional(),
      descricao: zod.string().optional(),
      confianca: zod.number(),
      nota: zod.string().optional(),
      balloonText: zod.string().optional(),
      imageThumbnail: zod.string().optional(),
      createdAt: zod.string().optional(),
      averageRating: zod.number().optional(),
      totalRatings: zod.number().optional(),
      searchType: zod.string().optional(),
      anoLancamento: zod.string().optional(),
      numeroPagina: zod.string().optional(),
      rank: zod.number().optional(),
    }),
  ),
  total: zod.number(),
  page: zod.number(),
  limit: zod.number(),
});

/**
 * @summary Get weekly ranking top 10
 */
export const GetRankingResponse = zod.object({
  items: zod.array(
    zod.object({
      id: zod.number().optional(),
      titulo: zod.string(),
      editora: zod.string(),
      imageThumbnail: zod.string().optional(),
      searchCount: zod.number(),
      averageRating: zod.number().optional(),
      totalRatings: zod.number().optional(),
    }),
  ),
  weekStart: zod.string(),
  weekEnd: zod.string(),
});

/**
 * @summary Submit a rating for a comic result
 */
export const SubmitRatingBody = zod.object({
  comicId: zod.number(),
  rating: zod
    .union([zod.literal(0), zod.literal(1)])
    .describe("0 = thumbs down, 1 = thumbs up"),
  feedback: zod.string().optional().describe("Text feedback when rating is thumbs down"),
});

export const SubmitRatingResponse = zod.object({
  success: zod.boolean(),
  averageRating: zod.number(),
  totalRatings: zod.number(),
});