# Downloads/gibifinder-flyio/gibifinder/lib/api-zod/src/generated/types/comicResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GibiFinder API - Comic book identifier using AI vision
 * OpenAPI spec version: 0.1.0
 */
import type { ComicResultSearchType } from "./comicResultSearchType";

export interface ComicResult {
  /** DB ID if saved */
  id?: number;
  encontrado: boolean;
  titulo: string;
  editora: string;
  personagens: string[];
  descricao: string;
  /**
   * @minimum 0
   * @maximum 100
   */
  confianca: number;
  nota: string;
  balloonText?: string;
  /** Base64 thumbnail of first image */
  imageThumbnail?: string;
  /** Average user rating (1=thumbs up, 0=thumbs down) */
  averageRating?: number;
  totalRatings?: number;
  searchType?: ComicResultSearchType;
}