# Downloads/gibifinder-flyio/gibifinder/lib/api-zod/src/generated/types/historyItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GibiFinder API - Comic book identifier using AI vision
 * OpenAPI spec version: 0.1.0
 */

export interface HistoryItem {
  id: number;
  titulo: string;
  editora: string;
  personagens?: string[];
  descricao?: string;
  confianca: number;
  nota?: string;
  balloonText?: string;
  imageThumbnail?: string;
  createdAt?: string;
  averageRating?: number;
  totalRatings?: number;
  searchType?: string;
}