# Downloads/gibifinder-flyio/gibifinder/lib/api-zod/src/generated/types/historyResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GibiFinder API - Comic book identifier using AI vision
 * OpenAPI spec version: 0.1.0
 */
import type { HistoryItem } from "./historyItem";

export interface HistoryResponse {
  items: HistoryItem[];
  total: number;
  page: number;
  limit: number;
}