# Downloads/gibifinder-flyio/gibifinder/lib/api-zod/src/generated/types/rankingResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GibiFinder API - Comic book identifier using AI vision
 * OpenAPI spec version: 0.1.0
 */
import type { RankingItem } from "./rankingItem";

export interface RankingResponse {
  items: RankingItem[];
  weekStart: string;
  weekEnd: string;
}