# Downloads/gibifinder-flyio/gibifinder/lib/api-zod/src/generated/types/ratingRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GibiFinder API - Comic book identifier using AI vision
 * OpenAPI spec version: 0.1.0
 */
import type { RatingRequestRating } from "./ratingRequestRating";

export interface RatingRequest {
  comicId: number;
  /** 0 = thumbs down, 1 = thumbs up */
  rating: RatingRequestRating;
}