# Downloads/gibifinder-flyio/gibifinder/lib/api-zod/src/generated/types/textSearchRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GibiFinder API - Comic book identifier using AI vision
 * OpenAPI spec version: 0.1.0
 */

export interface TextSearchRequest {
  /** Character name or comic title */
  query: string;
}